Rename AbstractToolbarAction of Form view to AbstractFormToolbarAction
