+++ src/Sulu/Bundle/AdminBundle/Resources/js/views/Form/index.js
@@ -1,7 +1,7 @@
 // @flow
 import Form from './Form';
 import toolbarActionRegistry from './registries/ToolbarActionRegistry';
-import AbstractToolbarAction from './toolbarActions/AbstractToolbarAction';
+import AbstractFormToolbarAction from './toolbarActions/AbstractFormToolbarAction';
 import DeleteToolbarAction from './toolbarActions/DeleteToolbarAction';
 import SaveWithPublishingToolbarAction from './toolbarActions/SaveWithPublishingToolbarAction';
 import SaveToolbarAction from './toolbarActions/SaveToolbarAction';
@@ -11,7 +11,7 @@ export default Form;
 
 export {
     toolbarActionRegistry,
-    AbstractToolbarAction,
+    AbstractFormToolbarAction,
     DeleteToolbarAction,
     SaveWithPublishingToolbarAction,
     SaveToolbarAction,
