+++ src/Sulu/Bundle/AdminBundle/Resources/js/views/Form/tests/registries/ToolbarActionRegistry.test.js
@@ -1,6 +1,6 @@
 // @flow
 import toolbarActionRegistry from '../../registries/ToolbarActionRegistry';
-import AbstractToolbarAction from '../../toolbarActions/AbstractToolbarAction';
+import AbstractFormToolbarAction from '../../toolbarActions/AbstractFormToolbarAction';
 
 jest.mock('../../../../services/Initializer', () => jest.fn());
 jest.mock('../../toolbarActions/DeleteToolbarAction', () => jest.fn());
@@ -13,7 +13,7 @@ beforeEach(() => {
 });
 
 test('Clear all toolbar actions', () => {
-    toolbarActionRegistry.add('test1', AbstractToolbarAction);
+    toolbarActionRegistry.add('test1', AbstractFormToolbarAction);
     expect(Object.keys(toolbarActionRegistry.toolbarActions)).toHaveLength(1);
 
     toolbarActionRegistry.clear();
@@ -21,16 +21,16 @@ test('Clear all toolbar actions', () => {
 });
 
 test('Add adapter', () => {
-    toolbarActionRegistry.add('test1', AbstractToolbarAction);
-    toolbarActionRegistry.add('test2', AbstractToolbarAction);
+    toolbarActionRegistry.add('test1', AbstractFormToolbarAction);
+    toolbarActionRegistry.add('test2', AbstractFormToolbarAction);
 
-    expect(toolbarActionRegistry.get('test1')).toBe(AbstractToolbarAction);
-    expect(toolbarActionRegistry.get('test2')).toBe(AbstractToolbarAction);
+    expect(toolbarActionRegistry.get('test1')).toBe(AbstractFormToolbarAction);
+    expect(toolbarActionRegistry.get('test2')).toBe(AbstractFormToolbarAction);
 });
 
 test('Add adapter with existing key should throw', () => {
-    toolbarActionRegistry.add('test1', AbstractToolbarAction);
-    expect(() => toolbarActionRegistry.add('test1', AbstractToolbarAction)).toThrow(/test1/);
+    toolbarActionRegistry.add('test1', AbstractFormToolbarAction);
+    expect(() => toolbarActionRegistry.add('test1', AbstractFormToolbarAction)).toThrow(/test1/);
 });
 
 test('Get adapter of not existing key', () => {
