+++ src/Sulu/Bundle/AdminBundle/Resources/js/views/index.js
@@ -1,6 +1,6 @@
 // @flow
 import Form, {
-    AbstractToolbarAction as AbstractFormToolbarAction,
+    AbstractFormToolbarAction as AbstractFormToolbarAction,
     toolbarActionRegistry as formToolbarActionRegistry,
 } from './Form';
 import List, {
